Create  color_identify
In Python, using OpenCV to achieve the following: input variables a, b, c, and d, where (a, b) and (c, d) are the coordinates of the upper left and lower right corners of a rectangle respectively. Output the HSV values of the part of the image enclosed by this rectangle and determine the color based on this.